# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prime Home LB Ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, price_asc, price_desc]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: onSale
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistStatus"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistStatus"

  # ─── ADMIN DASHBOARD ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, stock]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        stock: { type: integer }
        sku: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        featured: { type: boolean }
        bestSeller: { type: boolean }
        onSale: { type: boolean }
        visible: { type: boolean }
        salesCount: { type: integer }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, slug, price, stock]
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        salePrice: { type: number }
        stock: { type: integer }
        sku: { type: string }
        categoryId: { type: integer }
        images: { type: array, items: { type: string } }
        featured: { type: boolean }
        bestSeller: { type: boolean }
        onSale: { type: boolean }
        visible: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        stock: { type: integer }
        sku: { type: string }
        categoryId: { type: ["integer", "null"] }
        images: { type: array, items: { type: string } }
        featured: { type: boolean }
        bestSeller: { type: boolean }
        onSale: { type: boolean }
        visible: { type: boolean }
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    WishlistStatus:
      type: object
      required: [inWishlist]
      properties:
        inWishlist: { type: boolean }

    # ─── ADMIN ────────────────────────────────────────────────────────────────
...