# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  stock: number;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  images?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  onSale?: boolean;
  visible?: boolean;
  salesCount?: number;
  createdAt?: string;
}