# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  salePrice?: number;
  stock: number;
  sku?: string;
  categoryId?: number;
  images?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  onSale?: boolean;
  visible?: boolean;
}